// HNBACommissionSystem/src/app/shared/models/PIDSearch.ts
export interface IPIDSearch{

    PID_RECEIPT_NO : string;
    PID_RECEIPT_DATE: string;
    PID_CUSTOMER: string;
    PID_PROPOSAL_NO: string;
    PID_POLICY_NO: string;
    PID_RECEIPT_AMT: string;
    PID_TIME_SLAB: string;
    PID_CONFIRM_AMT: string;
    PID_CONFIRM_DATE: string;
    PID_RV_NO: string;
    PID_BAL_TYPE: string;
    PID_AGT_CODE: string;
    PID_CHEQUE_NO: string;
}